Menambahkan fungsi ketika data tidak tersedia di read.php(kategori)

+++ kategori/read.php
@@ -11,23 +11,36 @@ $totalRows = mysqli_num_rows($eksekusi);
 
 <h3>DAFTAR KATEGORI</h3>
 <a href="insert.php">Kategori</a>
-<table border="1">
-    <tr>
-        <td>NO.</td>
-        <td>NAMA KATEGORI</td>
-        <td>KET</td>
-        <td>AKSI</td>
-    </tr>
+<p></p>
+<form action="" method="get">
+    Cari Data : <input type="text" name="cari">
+    <button type="submit">Search</button>
+</form>
 
-    <?php $no = 1;
-    do {  ?>
+<!-- MENAMBAHKAN FUNGSI KETIKA DATA TIDAK TERSEDIA(KOSONG) -->
+<?php if ($totalRows > 0) { ?>
+
+    <table border="1">
         <tr>
-            <td><?php echo $no; ?></td>
-            <td><?php echo $row['namakategori']; ?></td>
-            <td><?php echo $row['ketkategori']; ?></td>
-            <td><a href="update.php?id=<?php echo $row['idkategori']; ?>"> Edit</a> | <a href="delete.php?id=<?php echo $row['idkategori']; ?>">Hapus</td>
+            <td>NO.</td>
+            <td>NAMA KATEGORI</td>
+            <td>KET</td>
+            <td>AKSI</td>
         </tr>
-    <?php
-        $no++;
-    } while ($row = mysqli_fetch_assoc($eksekusi)); ?>
-</table>
+
+        <?php $no = 1;
+        do {  ?>
+            <tr>
+                <td><?php echo $no; ?></td>
+                <td><?php echo $row['namakategori']; ?></td>
+                <td><?php echo $row['ketkategori']; ?></td>
+                <td><a href="update.php?id=<?php echo $row['idkategori']; ?>"> Edit</a> | <a href="delete.php?id=<?php echo $row['idkategori']; ?>">Hapus</td>
+            </tr>
+        <?php
+            $no++;
+        } while ($row = mysqli_fetch_assoc($eksekusi)); ?>
+    </table>
+<?php } else {
+    echo "Data tidak tersedia";
+}
+?>
